refactor(engineer_loop): extract meeting_decisions helpers (#1266)
mod.rs: 406 -> 311 ✓ under 400
meeting_decisions.rs: 110 LOC

All 498 engineer_loop tests pass.
